Floor and ceiling clipping changed to 32 bit.
The scaling value that causes wobbly walls/weird scaling walls/etc is fine. The math checks out. The problem is that the 32-bit value then gets shoved in to a 16-bit value and the loss of precision causes negative values/invalid accesses/overflows/etc. So now we're not crashing.

Was a widescale change as it turns out. But now I've typedef'd everything that uses it with a unique type. Hopefully it'll never get lost in translation again.