Big chunk of changes for widescreen, including selections.
Distance lighting picks the right value, and I've removed the divide while I was at it.

MAXSCREENWIDTH at this point is more of a hindrance. Need to delete those fixed-width buffers entirely.

V_ scaling routines are completely broken.

SIMD rendering overlaps the buffer when it shouldn't. Turned off for now.

I think that's it.