So what we have here is working ZDoom node support.
Ling suggested The Given as a test map for the renderer now that Planisphere 2 is basically working. Seemed like a good idea, until it turned out that it's built with ZDoom nodes. Cleanbox implemented using nothing but the ZDoom wiki article. Had to calculate the values it skipped, hit run, and it worked perfectly first time. Noice.

Need to test some other ZDoom node format maps now. But this will also definitely be a good one to test render optimisations with.

(Note that if you try to sync and run the map yourself, you will hit errors. I've not decided how I want to handle them just yet.)