My rasteriser now no longer deals with visplanes.
Still haven't deleted the entire visplane structure, or supporting code. But I have taken a huge hatchet to it. It's on the way out for sure now, just a few last bits of cleanup to do.

So. All that visplane waste. I've deleted the lookups, just allocate blindly. All I care about is creating new raster regions so that the plane renderer can handle it. SIGNIFICANT reduction in memory. The scratchpad only uses about 2 megs on Planisphere 2. Also, because there's no search and additional maintenance I save 6 milliseconds on my test area in the same map.

I expect the time to go down a bit more, at which point I'll post about it.