New "first launch" dashboard feature.
We allow the user to configure Rum and Raisin Doom natively inside the application (ie without the old setup.exe program) and introduce them to the Dashboard at the same time. After an intro page, we dump the user in to the options window (still need to clean that up some more) and once they're satisfied they can hit "Play!" and get on with it.