This is a massive intertwined commit that supports flats as walls and walls as flats.
Lots of ripping out of old arrays and integrating new ones. Both textures and flats have separate lookup arrays that point back to the master tables. Will be easier this way to keep vanilla compatibility.

Animation specails in the playsim have been updated as a result. This needs to move out to the renderer to be quite honest.