Palette lookup works properly on GLES now
Yeah, so depending on GLES version your highp floats could be half
floats instead. Which leads to a complete loss of precision breakdown
when you're trying to get palette entry 247 and higher.

So we solve this in two steps:
* Since 0-255 is being mapped in to 0-1, we bring that mapping back to
where we want it by dividing by 256 and multiplying by 255 (represented
by the 0.99609375 constant.
* Since that breaks precision elsewhere, we nudge the lookup by 1/1024.

Bit dodgy, but you've just gotta think like a half-precision floating
point to get to something precise enough for your needs.