More cleaning up for DRS.
Realised that I didn't need to modify the light table functions at all, all my fancy math did was make it revert to values that 320x200 would give you. But interestingly, distance falloff changes if the render target height is not (200*2^N). Will have to dig in to that at some point.