Square pixels are now supported! FOV slider implemented with bugs (sky, HUD weapons, diminished lighting)!
Also put in a small bit of groundwork for freelook. That'll come when I deal with those mentioned bugs.

Also snuck in an option to just quit the game when you press ALT+F4 or close on the window. Annoyed me when I was working on things, hard to split out as a separate commit, you get it all at once.