I FINALLY worked out conveyors.
I dedided that since all mobjs get a thinker and tick every frame even with a zero state, and that the first thing their think function does is apply velocity, then heck it rather than try to do an extra trymove I'll just add velocity to mobjs on a conveyor and let friction sort it out. so I started brute forcing a value to multiply the sector's scroll rate by. 10% was too much, 9% was too little, 9.375% seems to be about right after a bit more iteration.

and then I realised: hold the fuck up, I already worked this shit out 11 years ago.

https://www.doomworld.com/idgames/prefabs/floordmo

I divide the movement of a polyobject per frame by 10.7 and apply that to the rider's velocity. and 1/10.7 is.... 9.346%. I was never happy with that value in my zdoom script, it seemed just ever so slightly off but by that point I was tired of brute forcing it. so going the multiply way instead of divide finally resulted in a way better value.

Going to refactor a few things in this file in to their own thinkers, that big ass if branch is annoying me now.